test(symfony): implement deprecated publisherinterface
